requirements.txt
PyMuPDF==1.25.4
boto3==1.37.20
gspread==6.2.0
oauth2client==4.1.3
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
protobuf==5.29.4
pytesseract==0.3.13
requests==2.32.3
simple-salesforce==1.12.6
google-generativeai==0.8.4
numpy==1.26.4
colorama==0.4.6


